apneavercel/requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lyricsgenius==3.0.1
MarkupSafe==3.0.2
nltk==3.9.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
soupsieve==2.6
SQLAlchemy==2.0.36
textblob==0.18.0.post0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
